Add files via upload
hi lads best i could do with that behavorial state machine, i dont know if its much good or not. If any you guys want to edit and move few things on the diagram that done seem right try and fix it sure be great. these last two diagrams are not the best ones to do. 
cheers