TeamProject Main
The main version. Everyone should be using this version as a baseline. If you are making an update, do NOT upload here. Create a new branch, call it the date or something. MAKE SURE THERE IS NOT ALREADY AN UPDATE BRANCH IN EXISTENCE FIRST!!! Any new updates will go on this new branch. DO NOT OVERRIDE SOMEONE ELSE'S WORK!!! USE FACEBOOK TO MAKE SURE!!! Once a version is stable & complete, merge it with the main branch. All uploads should be part of TeamProject folder, not in zip format, not called anything else.